Added Files used for Project
This includes:
"deaths_on_eight_thousanders.csv" which is a csv containing climbers that have died
"List of Highest Mountains on Earth" which contains a bunch of very tall mountain data
"The merginator" is the code we used to merge those two data sets
"df_mtn_death" is the resulting data set, which includes the height of the mountain each climber died on (and other less important columns)
and finally "Mountain_Death Analysis" contains the data visualizations and analysis from which we concluded that mountain height and climber deaths are at the very least correlated. There's also a visualization detailing cause of death based on whether the mountain was above or below the median height.